Feature/suggestion meta data (#136)
* Add the initial setup for suggesting meta data.

* Only process form input on actual value change.

* Add suggestion handling to MetadataFieldsetComponent

Introduced a SuggestionFunction type to handle suggestions in MetadataFieldsetComponent. Integrated suggestion functionality in transaction-form.tsx, and enhanced form fields to update correctly based on suggestions.

* Format code for better readability

Improved the consistency of the codebase by adding spaces around braces and aligning object properties. This should enhance readability and maintainability, making it easier for developers to follow the code standards.